cloud_functions/iss_api_store_feedback/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
functions-framework==3.*
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.16.4
